Improve color blending of text gradients by interpolating in CIE space
Background
===

When computers represent colors, they make use of sRGB space, a gamma-encoded value that allows for higher resolution of darker values ^[1] ^[2]. The purpose of sRGB is to prioritize human visual perception, which follows a power law relationship -- the less stimuli there is, the greater any slight differences are percieved compared to a corresponding change with a strong stimuli

However, when processing color data, these sRGB values must be converted into a linear space in order to be correctly manipulated. Doing a basic interpolation on sRGB values can result in gradients being too dark or getting muddied by incorrect colors.

One solution is to work in the gamma-decoded linear version of sRGB ^[3]. This improves blending significantly and just requires a piecewise function, one part linear and the other part a power function.

---

The Java standard library actually gives us a much better solution though - do the interpolation in CIE XYZ space, a color space designed to match human perceptual differences along any euclidean distance within it as represented in a 3D space. This allows for gradients, alpha blending, image scaling, and antialiasing with a correct interpolation of luminosity, chroma, and hue. It should be used for most non-realtime color processing and manipulation purposes.

I've implemented this and it gives much better results - significantly improving readability for most of the examples

One thing I wasn't sure about with this PR was the purpose of the `RawUtil` class - is this just helper functions for future use? I've implemented the gradient function in both `Palette` and `RawUtil` but the `RawUtil` implementation never actually gets used it seems.

Further work could add a separate color formatting command to work in CIE LCH mode with H going either counterclockwise or clockwise, allowing for hue-aware gradients to give much better rainbow generation results than the current `&x` mode.

---

References

[1]: https://www.youtube.com/watch?v=LKnqECcg6Gw
[2]: https://blog.johnnovak.net/2016/09/21/what-every-coder-should-know-about-gamma/
[3]: https://en.wikipedia.org/wiki/SRGB#The_sRGB_transfer_function_(%22gamma%22)