persistent cache
added a cache manager to serialise compiled cache list

to minimise getOfflinePlayer() calls, the cache is generated by <uuid>.dat files in the playerdata director.

player names are only rechecked amd updated in the cache if .dat file has a different lastmodified date.

cache is loaded from disk is empty, and saved to disk on disable